+++ app/controllers/socify/users/omniauth_callbacks_controller.rb
@@ -1,0 +1,18 @@
+module Socify
+  module Users
+    class OmniauthCallbacksController < Devise::OmniauthCallbacksController
+      def google_oauth2
+          # You need to implement the method below in your model (e.g. app/models/user.rb)
+          @user = Socify::User.from_omniauth(request.env['omniauth.auth'])
+
+          if @user.persisted?
+            flash[:notice] = I18n.t 'devise.omniauth_callbacks.success', kind: 'Google'
+            sign_in_and_redirect @user, event: :authentication
+          else
+            session['devise.google_data'] = request.env['omniauth.auth'].except(:extra) # Removing extra as it can overflow some session stores
+            redirect_to new_user_registration_url, alert: @user.errors.full_messages.join("\n")
+          end
+      end
+    end
+  end
+end
